fix: Hot
Range is not Number
