+++ src/Traits/Fields/NumberTrait.php
@@ -12,6 +12,8 @@ trait NumberTrait
 
     public int|float $step = 1;
 
+    protected bool $stars = false;
+
     public function min(int|float $min): static
     {
         $this->min = $min;
@@ -31,5 +33,17 @@ trait NumberTrait
         $this->step = $step;
 
         return $this;
+    }
+
+    public function stars(): static
+    {
+        $this->stars = true;
+
+        return $this;
+    }
+
+    public function withStars(): bool
+    {
+        return $this->stars;
     }
 }
